requirements.txt
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
tavily-python>=0.3.3
langgraph>=0.1.4
langgraph-checkpoint-sqlite>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
sqlalchemy>=2.0.0
jinja2>=3.1.2